Split out recordSaveEvent
